+++ eResturant_web/app/admin.py
@@ -1,7 +1,10 @@
 from django.contrib import admin
 
 # Register your models here.
-from .models import userProfile, Booking, Table, Menu, MealOrder
+#from .models import userProfile, Booking, Table, Menu, MealOrder
+from app.models.user_model import userProfile
+from app.models.booking_model import Booking, Table
+from app.models.meal_model import Menu, MealOrder
 
 admin.site.register(userProfile)
 admin.site.register(Booking)
