+++ eResturant_web/app/forms/__init__.py
@@ -1,0 +1,2 @@
+from .user_forms import NewUserForm, ProfileForm
+from .booking_forms import BookingForm, editBookingForm
