+++ eResturant_web/app/forms/booking_forms.py
@@ -1,24 +1,7 @@
 from django import forms
-from django.contrib.auth.models import User
-from django.contrib.auth.forms import UserCreationForm
 from django.conf import settings
-from .models import userProfile, Booking, MealOrder
+from app.models.booking_model import Table, Booking
 from django.forms import ValidationError
-
-
-class NewUserForm(UserCreationForm):
-    email = forms.EmailField()
-
-    class Meta:
-        model = User
-        fields = ['username', 'email', 'password1', 'password2']
-
-
-class ProfileForm(forms.ModelForm):
-
-    class Meta:
-        model = userProfile
-        fields = ['First Name', 'Last Name']
 
 
 class BookingForm(forms.ModelForm):
@@ -65,9 +48,3 @@ class editBookingForm(forms.ModelForm):
             'bookingStartDateTime',
             'bookingEndDateTime'
         ]
-
-
-class MealOrderForm(forms.ModelForm):
-    class Meta:
-        model = MealOrder
-        fields = '__all__'
