+++ eResturant_web/app/forms/user_forms.py
@@ -1,0 +1,20 @@
+from django import forms
+from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
+from django.conf import settings
+from app.models.user_model import userProfile
+
+
+class NewUserForm(UserCreationForm):
+    email = forms.EmailField()
+
+    class Meta:
+        model = User
+        fields = ['username', 'email', 'password1', 'password2']
+
+
+class ProfileForm(forms.ModelForm):
+
+    class Meta:
+        model = userProfile
+        fields = ['First Name', 'Last Name']
