+++ eResturant_web/app/models/__init__.py
@@ -1,0 +1,3 @@
+from .booking_model import Table, Booking
+from .user_model import userProfile
+from .meal_model import Menu, calculateCost, MealOrder
