+++ eResturant_web/app/models/meal_model.py
@@ -4,36 +4,9 @@ from django.conf import settings
 from datetime import datetime, timedelta
 from django.core.exceptions import ValidationError
 from django.db.models.deletion import CASCADE
-# Create your models here.
-
-# login/signup
+from .booking_model import Table
 
 
-class userProfile(models.Model):
-    user = models.OneToOneField(User, on_delete=CASCADE)
-    fName = models.CharField(max_length=50, name="First Name")
-    lName = models.CharField(max_length=50, name="Last Name")
-    isStaff = models.BooleanField(default=False)
-###############
-
-
-# table booking
-class Table(models.Model):
-    maxCapacity = models.IntegerField()
-    isBooked = models.BooleanField(default=False)
-
-
-class Booking(models.Model):
-    user = models.ForeignKey(
-        userProfile, on_delete=models.CASCADE, default=1)
-    table = models.ForeignKey(Table, on_delete=models.CASCADE, default="")
-    bookingStartDateTime = models.DateTimeField(default=datetime.now)
-    bookingEndDateTime = models.DateTimeField(default=datetime.now)
-    guests = models.IntegerField(default=1)
-######################
-
-
-# meal ordering
 class Menu(models.Model):
     item_choices = (
         ("E", "Entree"),
