+++ eResturant_web/app/templates/app/bookings.html
@@ -1,0 +1,10 @@
+{% extends 'app/base.html' %}
+
+{% block content %}
+    <h1>Your booking for {{object.user.name}} is on {{ object.date }}</h1>
+    <p>Your booking starts at {{object.bookingStartDateTime.time}} and finishes on {{object.bookingEndDateTime.time}}</p>
+    <p>Number of Guests: {{object.guests}}</p>  
+
+
+
+{% endblock content %}
