+++ eResturant_web/app/templates/app/bookings_edit.html
@@ -1,0 +1,10 @@
+{% extends 'app/base.html' %}
+
+{% block content %}
+    <h2>Edit your Booking</h2>
+
+    <form method='POST'> {% csrf_token %}
+        {{ form.as_p }}
+        <input type='submit' value='Save' />
+    </form>
+{% endblock content %}
