+++ eResturant_web/app/views/loginSignup.py
@@ -1,12 +1,12 @@
 
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
-from app.forms import UserCreationForm
+
 from django.contrib.auth import login, authenticate
 from django.contrib import messages
 from django.utils.timezone import datetime
 from django.contrib.auth.forms import AuthenticationForm
-from app.forms import NewUserForm, ProfileForm
+from app.forms.user_forms import NewUserForm, ProfileForm
 from django.contrib.auth import login, logout
