+++ eResturant_web/app/views/mealOrdering.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 from django.http import HttpResponse
-from app.models import MealOrder
-from app.forms import MealOrderForm
+#from app.models import MealOrder
+#from app.forms import MealOrderForm
 
 # Create your views here.
 
@@ -11,11 +11,11 @@ def meal_order_view(request):
     return render(request, 'app/meal_order.html', context)
 
 
-def make_order_view(request):
+"""def make_order_view(request):
 
     form_class = MealOrderForm
     form = form_class(request.POST or None)
 
     context = {'form': form}
 
-    return render(request, 'app/order.html', context)
+    return render(request, 'app/order.html', context)"""
