--- eResturant_web/templates/booking_create.html
@@ -1,8 +1,0 @@
-{% extends 'app/base.html' %}
-
-{% block content %}
-    <form method='POST'> {% csrf_token %}
-        {{ form.as_p }}
-        <input type='submit' value='Save' />
-    </form>
-{% endblock content %}
